[IMP] l10n_in: clean by removing tax splitting by hsn
There is no need anymore to split tax lines by hsn as
the report now uses the tax details to calculate the
taxes again anyways.

And we can group the reverse tax fields on tax and
tax template in one file then.
